Sheet#1

+++ #1/D.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    int a,b,c,d;
+    long long a,b,c,d;
     cin >> a >> b >> c >> d;
     cout << "Difference = " << (a*b) - (c*d);
 }
